В метод удаления клиента добавлена проверка на то, что у клиента нет не погашенных кредитов. При удалении клиенту удаляются так же все связанные аккаунты и кредиты.

+++ src/main/java/bank/application/service/ClientServiceImpl.java
@@ -28,6 +28,7 @@ public class ClientServiceImpl implements ClientService {
     private CreditDao creditDao;
 
 
+    // в методе осуществляется проверка что клиенту больше 18 лет
     @Override
     public Client addNewClient(final String fio, final int age) throws ClientException {
         if (age >= CLIENT_MIN_AGE) {
@@ -48,11 +49,36 @@ public class ClientServiceImpl implements ClientService {
         throw new ClientException("Клиент не найден.");
     }
 
+    // в методе осуществляется проверка то, что у клиента нет не погашенных кредитов
+    // нельзя удалить клиента с непогашенными кредитами
+    // при удалении клиента, так же удаляются все свзанные аккаунты и кредиты
     @Override
     public void deleteClient(final Integer clientId) throws ClientException {
         Client client = clientDao.getClientById(clientId).orElse(null);
+        ArrayList<Credit> clientCredits = (ArrayList<Credit>) getAllClientCredits(clientId);
+        ArrayList<Credit> clientNotPaidCredits = new ArrayList<>();
+        if (clientCredits.size() > 0) {
+            for (int i = 0; i < clientCredits.size(); i++) {
+                if (!clientCredits.get(i).isPaidOff()) {
+                    clientNotPaidCredits.add(clientCredits.get(i));
+                }
+            }
+        }
         if (client != null) {
-            clientDao.deleteClient(clientId);
+            if (clientNotPaidCredits.size() == 0) {
+                clientDao.deleteClient(clientId);
+                ArrayList<Account> clientAccounts = (ArrayList<Account>) getAllClientAccounts(clientId);
+                for (int i = 0; i < clientAccounts.size(); i++) {
+                    Integer accountId = clientAccounts.get(i).getAccountId();
+                    accountDao.deleteAccount(accountId);
+                }
+                for (int i = 0; i < clientCredits.size(); i++) {
+                    Integer creditId = clientCredits.get(i).getCreditId();
+                    creditDao.deleteCredit(creditId);
+                }
+            } else {
+                throw new ClientException("Клиента нельзя удалить. У клиента имеются непогашенные кредиты.");
+            }
         } else {
             throw new ClientException("Клиент не найден.");
         }
